+++ app/helpers/application_helper.rb
@@ -4,16 +4,7 @@ module ApplicationHelper
     content_for(:title) { page_title }
   end
 
-  def valid_year
-    Time.last_year.to_s
-  end
-
-  def current_year_voluntary_info?(ori)
-    AdditionalSubmissionInformation.where(submission_year: valid_year)
-                                   .any? { |asi| asi.ori == ori }
-  end
-
-  def class_names
+  def body_class
     [controller_name, action_name].map(&:parameterize).join(' ')
   end
 end
