+++ app/views/layouts/application.html.erb
@@ -1,7 +1,7 @@
 <!DOCTYPE html>
 <html>
 <%= render "layouts/partials/head" %>
-<body class="<%= class_names %>">
+<body class="<%= body_class %>">
   <div id="container">
     <%= render "layouts/partials/nav" %>
     <div id="body">
